Remove `TokenData` Ownership from `Parser`
This refactor was motivated by 2 points
* the Salsa library by Nikos Matsakis
* the talk "the life of a file" by Evan Czaplicki
    where he recommends thinking about who owns state.
    The Rust lanaguage promotes this thinking anyways

The goal was to change the ownership of all the tokens from
the Parser to some auxilary data-structure under the AST.
Instead of creating a new data-structure, I use a reference-counted
point as is common in the Salsa library. This change, as all good
ones do, lead me to simplify and remove chunks of code to manage the
parser and lexer state.

TODO:
* testing - need to lay foundation for extensive compiler testing
    framework.
* parser error recovery